pyproject.toml
[project]
name = "ppt-agent"
version = "0.1.0"
description = "AI-powered PPT generation agent for business presentations"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.32.0",
    "anthropic>=0.25.0",
    "python-pptx>=0.6.23",
    "pydantic>=2.6.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
